pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "STXit"
version = "1.0.3"
description = "STX subtype calling and prophage-context reporting for assembled bacterial genomes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "biopython>=1.80",
  "pandas>=2.0",
  "matplotlib>=3.8",
  "requests>=2.31"
]

[project.scripts]
stxit = "stxit.cli:main"

[tool.setuptools.packages.find]
include = ["stxit*"]
